now user's credited and debited amount from choosen account  will be displayed on the account section, live data update is achieved

+++ app/src/main/java/com/lazysloth/pocketlog/database/data/Transaction.kt
@@ -23,7 +23,7 @@ data class Transaction(
     val id: Int = 0,
     val userId: Int?,
     val amount: Double,
-    val accountId: Int,
+    val accountId: Long,
     val category: Category,
     val transactionType: TransactionType,
     val note: String,
